+++ packages/blockstore-idb/src/index.ts
@@ -15,6 +15,8 @@
 import { BaseBlockstore } from 'blockstore-core'
 import { openDB, deleteDB } from 'idb'
 import { OpenFailedError, PutFailedError, NotFoundError } from 'interface-store'
+import all from 'it-all'
+import toBuffer from 'it-to-buffer'
 import { base32upper } from 'multiformats/bases/base32'
 import { CID } from 'multiformats/cid'
 import * as raw from 'multiformats/codecs/raw'
@@ -22,7 +24,7 @@ import * as Digest from 'multiformats/hashes/digest'
 import { raceSignal } from 'race-signal'
 import type { IDBPDatabase } from 'idb'
 import type { Pair } from 'interface-blockstore'
-import type { AbortOptions, AwaitIterable } from 'interface-store'
+import type { AbortOptions, AwaitGenerator, AwaitIterable } from 'interface-store'
 import type { MultibaseCodec } from 'multiformats/bases/interface'
 
 export interface IDBBlockstoreInit {
@@ -85,14 +87,22 @@ export class IDBBlockstore extends BaseBlockstore {
     this.db?.close()
   }
 
-  async put (key: CID, val: Uint8Array, options?: AbortOptions): Promise<CID> {
+  async put (key: CID, val: Uint8Array | AwaitIterable<Uint8Array>, options?: AbortOptions): Promise<CID> {
     if (this.db == null) {
       throw new Error('Blockstore needs to be opened.')
     }
 
+    let buf: Uint8Array
+
+    if (val instanceof Uint8Array) {
+      buf = val
+    } else {
+      buf = toBuffer(await all(val))
+    }
+
     try {
       options?.signal?.throwIfAborted()
-      await raceSignal(this.db.put(this.location, val, this.#encode(key)), options?.signal)
+      await raceSignal(this.db.put(this.location, buf, this.#encode(key)), options?.signal)
     } catch (err: any) {
       throw new PutFailedError(String(err))
     }
@@ -100,7 +110,7 @@ export class IDBBlockstore extends BaseBlockstore {
     return key
   }
 
-  async get (key: CID, options?: AbortOptions): Promise<Uint8Array> {
+  async * get (key: CID, options?: AbortOptions): AwaitGenerator<Uint8Array> {
     if (this.db == null) {
       throw new Error('Blockstore needs to be opened.')
     }
@@ -118,7 +128,7 @@ export class IDBBlockstore extends BaseBlockstore {
       throw new NotFoundError()
     }
 
-    return val
+    yield val
   }
 
   async delete (key: CID, options?: AbortOptions): Promise<void> {
@@ -149,7 +159,7 @@ export class IDBBlockstore extends BaseBlockstore {
     }
   }
 
-  async * getAll (options?: AbortOptions): AwaitIterable<Pair> {
+  async * getAll (options?: AbortOptions): AwaitGenerator<Pair> {
     if (this.db == null) {
       throw new Error('Blockstore needs to be opened.')
     }
@@ -159,9 +169,11 @@ export class IDBBlockstore extends BaseBlockstore {
     for (const key of await this.db.getAllKeys(this.location)) {
       options?.signal?.throwIfAborted()
       const cid = this.#decode(key.toString())
-      const block = await this.get(cid, options)
 
-      yield { cid, block }
+      yield {
+        cid,
+        bytes: this.get(cid, options)
+      }
 
       options?.signal?.throwIfAborted()
     }